refactor: message가 커짐에 따라 모듈 생성 후 구분#188
- Message repository 가 단순히 채널에 국한 되지 않고, User 와의 관계도
  갖게 되어 모듈을 생성했고, 외부로 꺼냈다. 때문에 AppModule 에도 import 를
진행하였다.

* channels.service
- Message Module 이 생성 되고, 해당 서비스에서 messageRepository 를 직접 사용하는 것을 발견 하여 이를 service를 호출하여 처리하는 것으로 변경함
