+++ back/web-app/src/message/entity/message.entity.ts
@@ -1,5 +1,5 @@
-import { Column, CreateDateColumn, Entity, JoinColumn, ManyToOne, PrimaryGeneratedColumn } from "typeorm";
-import { Channel } from "./channel.entity";
+import { Column, Entity, JoinColumn, ManyToOne, PrimaryGeneratedColumn } from "typeorm";
+import { Channel } from "../../channels/channel/entity/channel.entity";
 
 @Entity('message')
 export class Message {
